[fix]イベント削除機能の修正
イベント削除機能の修正
